+++ src/components/Hooks/Paging.js
@@ -1,5 +1,5 @@
 import Pagination from "react-js-pagination";
-import Pagintion from "../components/styles/Pagination.css";
+import Pagintion from "../styles/Pagination.css";
 
 const Paging = ({ page, count, data, handlePageChange }) => {
   console.log(page, data);
